api/requirements.txt
fastapi>=0.115.11
uvicorn[standard]==0.34.0
anthropic==0.43.0
asyncpg>=0.29.0
python-dotenv==1.0.1
rank-bm25==0.2.2
pydantic==2.10.4
google-genai==1.14.0
httpx==0.28.1
slowapi==0.1.9
PyJWT>=2.8.0
notebooklm-py>=0.3.3
sentry-sdk[fastapi]>=2.0.0
yfinance>=0.2.54
PyMuPDF>=1.24.0
python-docx>=1.1.0
python-multipart>=0.0.9
